requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==6.0.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
boto3==1.38.24
botocore==1.38.24
Brotli==1.1.0
browser-use==0.2.5
cachetools==5.5.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
courlan==1.3.2
cryptography==46.0.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.3.1
dateparser==1.2.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
duckduckgo_search==8.0.2
e2b==1.5.0
ensure==1.0.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.6
googleapis-common-protos==1.67.0
gradio==5.27.0
gradio_client==1.9.0
greenlet==3.2.2
groovy==0.1.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
hyperframe==6.1.0
ibm-cos-sdk==2.13.5
ibm-cos-sdk-core==2.13.5
ibm-cos-sdk-s3transfer==2.13.5
ibm_watsonx_ai==1.3.23
idna==3.10
inflate64==1.0.1
instaloader==4.15
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
json_repair==0.46.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jusText==3.0.2
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-community==0.3.20
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ibm==0.3.10
langchain-mcp-adapters==0.0.9
langchain-mistralai==0.2.4
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.34
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.70
langsmith==0.3.8
lomond==0.3.3
lxml==5.3.1
lxml_html_clean==0.4.2
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mediapipe==0.10.31
mem0ai==0.1.93
mistune==3.1.1
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
mypy-boto3-s3==1.36.21
mypy_extensions==1.1.0
narwhals==2.16.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.6
ollama==0.4.9
openai==1.82.0
opencv-python==4.12.0.88
orjson==3.10.15
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patchright==1.52.4
pillow==10.2.0
platformdirs==4.3.6
playwright==1.52.0
portalocker==2.10.1
postgrest==2.28.0
posthog==3.25.0
primp==0.15.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py7zr==0.22.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.3
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.12.5
pydantic-settings==2.9.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.2.1
pyperclip==1.9.0
pyppmd==1.1.1
pyroaring==1.0.3
python-box==6.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
pyzstd==0.16.2
qdrant-client==1.14.2
realtime==2.28.0
redis==6.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
ruff==0.11.11
s3transfer==0.13.0
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
screeninfo==0.8.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==70.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
smolagents==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.6
SQLAlchemy==2.0.41
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
storage3==2.28.0
streamlit==1.54.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
-e git+https://github.com/krishnaik06/Text-Summarization-NLP-Project.git@330811229b1a3db4674446e4984090e8afb19093#egg=textSummarizer
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tld==0.13.1
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1+cu118
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.49.0
twilio==9.10.2
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uuid7==0.1.0
uvicorn==0.34.2
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.22.0
yt-dlp==2026.2.4
zstandard==0.23.0
